[innovation-week-25q1-custom-span] Added and fixed tests
